+++ internal/data/fixtures.go
@@ -511,9 +511,9 @@ func CreatePaymentFixture(t *testing.T, ctx context.Context, sqlExec db.SQLExecu
 	const query = `
 		INSERT INTO payments
 			(receiver_id, disbursement_id, receiver_wallet_id, asset_id, amount, status, status_history,
-			stellar_transaction_id, stellar_operation_id, created_at, updated_at, external_payment_id, type)
+			stellar_transaction_id, stellar_operation_id, created_at, updated_at, external_payment_id, type, sender_address)
 		VALUES
-			($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13)
+			($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14)
 		RETURNING
 			id
 	`
@@ -532,6 +532,7 @@ func CreatePaymentFixture(t *testing.T, ctx context.Context, sqlExec db.SQLExecu
 		p.UpdatedAt,
 		p.ExternalPaymentID,
 		p.Type,
+		p.SenderAddress,
 	)
 	require.NoError(t, err)